PCB Rev 2: Create basic board outline
Re-imported the vector drawing of the mechanical board outline and put
the mirrored copy on a different layer so it can be differentiated by
color.
Eco1.User is the unmirrored drawing.
Eco2.User is the mirrored drawing

Added some mechanical keep-out areas to these layers as well.
I probably need two new layers for these but, for now, they are on the
same user layers.
The wide circles on Eco1.User are the keep-out areas behind the d-pad;
so these will actually collide with components placed on the back side
of the board.
The wide circles on Eco2.User are the keep-out areas behind the ABXY
buttons; so these will actually collide with components placed on the
front side of the board.

Created a board outline only for the main section.
I think I will need the extra room and have to use all 3 of these
for the main circuit board. The LR button boards can be done separately
if desired; but, for now, there isn't a good place to stick them
without increasing board rectangular size substantially.
Also OSH Park doesn't support V-grooves.