Schematic Rev 3 - Rework board-to-board jumpers
Decided to go with 3 full size boards to fill the mechanical space in
the controller instead of using the 3rd board as a spare.
This allows to move CN1 and both ICs onto the center board, which now
has plenty of space.
This solves several problems working around plastic ridges behind the
boards where I wanted to place components and reduces the number of
board-to-board jumpers needed.