PCB: Move TR/TL Connectors to avoid fouling on CN1
These connectors were also swapped to the wrong side of the board
which is partly why I didn't notice the mechanical collision.
I'm debating whether they need to actually be moved to the DPAD
and ABXY boards, but that is a pretty big tear up.
Anyway, I took the opportunity to optimize some of the GND routing.