Route TR/TL through the board-to-board connectors
Having TR and TL connectors on the main board puts them too far
away from the trigger boards. This prevents the existing wires
on those boards from being reused.

Move the connectors to the center of the top edge of the ABXY
and DPAD boards (a single connector can be used on both boards).
This combines P11 and P13 (now P15).

P15 is routed straight to the board-to-board connectors so that
it can be brought onto the main board. This changes these from
1x5 to 1x6 pin headers. I have changed the designator for these:
P4 -> P14
P6 -> P16

I have also changed the board-to-board connectors to 2.54mm
pitch because the 1.27mm pitch was extremely tedious to hand-
solder.